+++ src/components/todos/todos-section.tsx
@@ -1,72 +1,18 @@
-import { todoApi } from '../../api/todoApi'
-import type { Todo } from '../../types'
+import { useTodos } from '../../hooks/useTodos'
 import { Spinner } from '../spinner'
 import { TodoForm } from './todo-form'
 import { TodoItem } from './todo-item'
-import { useEffect, useState } from 'react'
+import { ErrorMessage } from '../error-message'
 
 export const TodosSection = () => {
-  const [todos, setTodos] = useState<Todo[]>([])
-  const [isLoading, setIsLoading] = useState(false)
-
-  const fetchTodos = async () => {
-    setIsLoading(true)
-    try {
-      const data = await todoApi.fetchTodos()
-      setTodos(data)
-    } catch (error) {
-      console.error('Failed to fetch todo tasks:', error)
-    } finally {
-      setIsLoading(false)
-    }
-  }
-
-  const addTodo = async (todoName: string) => {
-    setIsLoading(true)
-    try {
-      const newTodo = await todoApi.createTodo(todoName)
-      setTodos((prevTodos) => {
-        return [...prevTodos, newTodo]
-      })
-    } catch (error) {
-      console.error(error)
-    } finally {
-      setIsLoading(false)
-    }
-  }
-  const deleteTodo = async (todoId: number) => {
-    setIsLoading(true)
-    try {
-      await todoApi.deleteTodo(todoId)
-      setTodos((prevTodos) => prevTodos.filter((todo) => todo.id !== todoId))
-    } catch (error) {
-      console.error(error)
-    } finally {
-      setIsLoading(false)
-    }
-  }
-
-  const toggleTodo = async (todoId: number, completed: boolean) => {
-    setIsLoading(true)
-    try {
-      const updatedTodo = await todoApi.toggleTodo(todoId, !completed)
-      setTodos((prevTodos) => prevTodos.map((todo) => (todo.id === todoId ? updatedTodo : todo)))
-    } catch (error) {
-      console.error(error)
-    } finally {
-      setIsLoading(false)
-    }
-  }
-
-  useEffect(() => {
-    fetchTodos()
-  }, [])
+  const { addTodo, isLoading, todos, deleteTodo, toggleTodo, error, refetch } = useTodos()
 
   return (
     <main>
+      {error && <ErrorMessage message={error} onDismiss={refetch} />}
       <TodoForm addTodo={addTodo} />
       <div className="todo-container">
-        <ul id="todo-list">
+        <ul id="todo-list" className={isLoading === true ? 'isLoading' : ''}>
           {todos.map((todo) => {
             return <TodoItem key={todo.id} todo={todo} deleteTodo={deleteTodo} toggleTodo={toggleTodo} />
           })}
